requirements.txt
appdirs==1.4.3
appnope==0.1.0
autopep8==1.3.1
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
certifi==2017.1.23
decorator==4.0.11
Django==1.8
django-bulk-update==1.1.10
django-graphos==0.3.41
django-tables2==1.4.2
enum34==1.1.6
ipython-genutils==0.2.0
numpy==1.12.1
packaging==16.8
pathlib2==2.2.1
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.14
ptyprocess==0.5.1
pycodestyle==2.3.1
pycurl==7.43.0
Pygments==2.2.0
pynsq==0.8.0
pyparsing==2.2.0
requests==2.13.0
scandir==1.5
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
tornado==4.4.3
traitlets==4.3.2
ujson==1.35
wcwidth==0.1.7
